refactor: entities imports
