+++ packages/deployment-service/src/core/index.ts
@@ -1,2 +1,1 @@
 export * from './db'
-export * from './logger'
