+++ packages/deployment-service/src/functions/index.ts
@@ -2,3 +2,4 @@ export * from './release'
 export * from './pipeline'
 export * from './pusher-auth'
 export * from './migrationRun'
+export * from './github-webhook'
