+++ packages/deployment-service/src/functions/pipeline/pipeline-create.ts
@@ -1,6 +1,6 @@
 import { BadRequestException, httpHandler, ValidationException } from '@homeservenow/serverless-aws-handler'
 import { PipelineDto } from './../../dto'
-import { PipelineEntity } from './../../entities'
+import { PipelineEntity } from '../../entities/pipeline.entity'
 import * as service from './../../services/pipeline'
 import { plainToClass } from 'class-transformer'
 import { validate } from 'class-validator'
