+++ packages/deployment-service/src/services/index.ts
@@ -4,3 +4,4 @@ export * from './sqs'
 export * from './task'
 export * from './s3'
 export * from './pusher'
+export * from './github-app'
