+++ packages/deployment-service/src/utils/project-dir.ts
@@ -1,5 +1,5 @@
 import { resolve } from 'path'
-import { PipelineEntity } from './../entities'
+import { PipelineEntity } from '../entities/pipeline.entity'
 
 export const projectDir = (dir: string, pipeline: PipelineEntity) =>
   resolve(dir, `${(pipeline.repository as string).split('/').pop()}-master`)
